Refactored EntityClass
Its original intention was to distinguish unrenderable entities from
renderables. But the UI system revealed that it's real use is to
identify entities loosely associated with specific render or update
passes.
However, not every entity of a pass needs to be rendered. Hence,
this distinction was moved to the `isrenderable` method.